first hack at tree viz
